+++ ricaun.RevitTest.TestAdapter/Services/RevitTestConsole.cs
@@ -109,12 +109,12 @@ namespace ricaun.RevitTest.TestAdapter.Services
             string language = null,
             bool revitOpen = false,
             bool revitClose = false,
+            int timeoutMinutes = 0,
             Action<string> consoleAction = null,
             Action<string> debugAction = null,
             Action<string> errorAction = null,
             params string[] filter)
         {
-            var timeoutMinutes = 0;
             await ricaun.RevitTest.Command.Process.RevitTestProcessStartUtils.RunExecuteTests(
                 applicationPath, file, testResultAction,
                 version, language, revitOpen, revitClose, timeoutMinutes,